pyproject.toml
[tool.poetry]
name = "imagecrud"
version = "0.1.0"
description = ""
authors = ["reezuleanu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.7"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
pydantic = "^2.8.2"
mongoengine = "^0.29.0"
python-multipart = "^0.0.9"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
